fix: try fix CodeQL scanning

+++ jadx-gui/src/main/java/jadx/gui/ui/hexviewer/HexPreviewPanel.java
@@ -291,7 +291,7 @@ public class HexPreviewPanel extends JPanel {
 
 			char[] targetData = new char[textLength];
 			Arrays.fill(targetData, ' ');
-			for (int i = 0; i < copy.getDataSize(); i++) {
+			for (int i = 0; i < (int) copy.getDataSize(); i++) {
 				CodeAreaUtils.byteToCharsCode(copy.getByte(i), codeType, targetData, i * charsPerByte, charactersCase);
 			}
 			return new String(targetData);
